chore: Move cell magic code into its own directory
