+++ docs/magics.rst
@@ -1,5 +1,5 @@
 IPython Magics for BigQuery
 ===========================
 
-.. automodule:: google.cloud.bigquery.magics
+.. automodule:: google.cloud.bigquery.ipython_magics.magics
     :members:
