+++ google/cloud/bigquery/ipython_magics/magics.py
